+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # Finnick's Changelog
 Listed below are the changes made to the Finnick bot.
 
+## Version 0.0.2 
+
+### Added
+
+- Command Interface
+
+### Changed
+
+- Instead of Mentioning the path of where is FClient you can do '@lib/FClient' instead!
+
 ## Version 0.0.1 - 12/7/20
 
 ### Added